add IndRunner to help with induction equiv proofs
I intended to use these results in a proof of a forking lemma with
a standard RO (instead of FRO), but have since then came up with
a much simpler solution.

As a result, the added lemmas likely won't be used in the next
couple of commits, but for now, I want to keep them around, should
they become useful later.