add input generators to Runner and Forker
Runner and Forker now have variants, IRunner and IForker, that
take an input generator module as an argument.

This is an important addition for using the forking lemma in
security games --- reductions usually work with a fixed input
and hence need the basic Forker, but to analyse the probability
of a success event relative to the whole game, we need to average
over different inputs generated by the game and in that case, the
IForker variant becomes useful.

(The diff is quite large, but there's nothing too fancy going on,
it's mostly renaming, changing module and procedure arguments, and
adding the assertions that hops preserve equality of globs of the
input generator I.)